refactor: merge v1/v2 tests into a single compilation unit
mainly because this merges their dead code evaluation, so `common` no
longer has "unused functions".

also they are literally the same test cases. why were they separated lol
